Comment out edit why PM button in profile
